items: type EconomicValue as the float stock parses
ItemClass.EconomicValue is a Single parsed with ParseFloat (IL_0666).
zdtd typed it u16, so a modlet price above 65535 (or a decimal) parsed to
0 - the trader's untradeable marker - and the item was stocked at the
buy-5 / sell-1 fallback and rejected by traderSellPrice instead of trading
at its authored price.

ItemDef.econ and the item_modifiers row are f32 now; the price formulas
(which already computed in f64) are unchanged. A test pins 1000000 and
2.5 round-tripping.

+++ src/server/game/trader.zig
@@ -221,7 +221,7 @@ pub fn fillTraderFromXml(self: *Game, trader_net_id: i32) void {
         if (n >= ecs.components.max_stock) break;
         const iid = self.ecsIdFromItemName(r.name);
         if (iid == 0) continue;
-        const econ: u16 = if (self.items.byId(iid)) |d| d.econ else 0;
+        const econ: f32 = if (self.items.byId(iid)) |d| d.econ else 0;
         // A39: the sell base is EconomicValue * EconomicSellScale (stock
         // GetSellPrice; default scale 1.0, a few items mark down to .5).
         const sell_scale: f32 = if (self.items.byId(iid)) |d| d.econ_sell_scale else 1.0;
